+++ pdf-converter/CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## 1.0.1
+## Added
+- add keyup handler for dialog
+- new description for admin settings
+ 
 ## 1.0.0
 ## Added
 - first release